requirements.txt
Django==2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
google-api-python-client==1.8.0
google-auth-oauthlib==0.4.1

